requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
